refactor: CourseService SRP 분리 및 트랜잭션 정합성 개선
- AiCourseSaveService, CourseBakeryOrderService, CourseDrivingRouteService 신규 분리
- CourseThumbnailAssembler 추출
- getDrivingRoute readOnly = true 제거
