+++ apps/be/src/test/java/com/breadbread/course/service/CourseServiceTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.never;
@@ -12,34 +11,17 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.breadbread.bakery.entity.Bakery;
-import com.breadbread.bakery.entity.BakeryImage;
 import com.breadbread.bakery.entity.BreadType;
-import com.breadbread.bakery.repository.BakeryImageRepository;
 import com.breadbread.bakery.repository.BakeryRepository;
-import com.breadbread.bakery.service.BakeryImageUrlResolver;
-import com.breadbread.course.client.DrivingRouteClient;
-import com.breadbread.course.dto.Coordinate;
 import com.breadbread.course.dto.CourseListResponse;
 import com.breadbread.course.dto.CourseSearch;
-import com.breadbread.course.dto.DrivingRouteResponse;
 import com.breadbread.course.dto.ManualCourseRequest;
-import com.breadbread.course.dto.ReorderBakeriesRequest;
-import com.breadbread.course.dto.ReorderBakeriesResponse;
 import com.breadbread.course.dto.RouteResponse;
-import com.breadbread.course.dto.RouteResult;
 import com.breadbread.course.dto.UpdateCourseRequest;
-import com.breadbread.course.dto.ai.AiCoursePreviewResponse;
-import com.breadbread.course.dto.ai.AiCourseRequest;
-import com.breadbread.course.dto.ai.AiCourseResultCache;
-import com.breadbread.course.dto.ai.AiCourseWebhookResponse;
-import com.breadbread.course.dto.ai.AiJobStatus;
-import com.breadbread.course.dto.ai.AiJobStatusResponse;
-import com.breadbread.course.dto.ai.RecommendedBakeryResponse;
 import com.breadbread.course.entity.AiCourseInfo;
 import com.breadbread.course.entity.BudgetRange;
 import com.breadbread.course.entity.Course;
 import com.breadbread.course.entity.CourseBakery;
-import com.breadbread.course.entity.CourseDrivingRoute;
 import com.breadbread.course.entity.CourseLike;
 import com.breadbread.course.entity.FlexibilityLevel;
 import com.breadbread.course.entity.ManualCourseInfo;
@@ -50,23 +32,17 @@ import com.breadbread.course.repository.CourseDrivingRouteRepository;
 import com.breadbread.course.repository.CourseLikeRepository;
 import com.breadbread.course.repository.CourseRepository;
 import com.breadbread.course.repository.RouteRepository;
-import com.breadbread.course.service.ai.AiCourseAsyncService;
-import com.breadbread.course.service.ai.AiCourseRedisService;
-import com.breadbread.course.service.ai.AiCourseResultRedisService;
 import com.breadbread.global.exception.CustomException;
 import com.breadbread.global.exception.ErrorCode;
 import com.breadbread.user.entity.User;
-import com.breadbread.user.entity.UserPreference;
 import com.breadbread.user.entity.UserRole;
-import com.breadbread.user.entity.WaitingTolerance;
-import com.breadbread.user.repository.UserPreferenceRepository;
 import com.breadbread.user.repository.UserRepository;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.CompletableFuture;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
@@ -85,19 +61,11 @@ class CourseServiceTest {
     @Mock private CourseRepository courseRepository;
     @Mock private BakeryRepository bakeryRepository;
     @Mock private CourseBakeryRepository courseBakeryRepository;
-    @Mock private BakeryImageRepository bakeryImageRepository;
     @Mock private CourseLikeRepository courseLikeRepository;
     @Mock private UserRepository userRepository;
     @Mock private RouteRepository routeRepository;
     @Mock private CourseDrivingRouteRepository courseDrivingRouteRepository;
-    @Mock private AiCourseAsyncService aiCourseAsyncService;
-    @Mock private AiCourseRedisService aiCourseRedisService;
-    @Mock private AiCourseResultRedisService aiCourseResultRedisService;
-    @Mock private UserPreferenceRepository userPreferenceRepository;
-    @Mock private DrivingRouteClient drivingRouteClient;
-    @Mock private CourseDrivingRouteSaver courseDrivingRouteSaver;
-    @Mock private BakeryImageUrlResolver bakeryImageUrlResolver;
-    @Mock private com.breadbread.tour.service.TourRedisService tourRedisService;
+    @Mock private CourseThumbnailAssembler courseSummaryAssembler;
 
     @InjectMocks private CourseService courseService;
 
@@ -111,16 +79,8 @@ class CourseServiceTest {
         Pageable pageable = PageRequest.of(0, 10);
         when(courseRepository.search(any(CourseSearch.class), eq(pageable)))
                 .thenReturn(new PageImpl<>(List.of(course), pageable, 1));
-        when(bakeryImageRepository.findAllByBakeryIdInAndDisplayOrder(List.of(10L), 1))
-                .thenReturn(
-                        List.of(
-                                BakeryImage.builder()
-                                        .imageUrl("thumb.jpg")
-                                        .displayOrder(1)
-                                        .bakery(bakery)
-                                        .build()));
-        when(bakeryImageUrlResolver.resolve(any(BakeryImage.class)))
-                .thenAnswer(inv -> ((BakeryImage) inv.getArgument(0)).getImageUrl());
+        when(courseSummaryAssembler.buildThumbnailMap(List.of(10L)))
+                .thenReturn(Map.of(10L, "thumb.jpg"));
         when(courseLikeRepository.countByCourseIdIn(List.of(1L)))
                 .thenReturn(Collections.singletonList(new Object[] {1L, 3L}));
         when(courseLikeRepository.findLikedCourseIdsByUserId(List.of(1L), 99L))
@@ -145,8 +105,6 @@ class CourseServiceTest {
         Pageable pageable = PageRequest.of(0, 5);
         when(courseRepository.search(any(CourseSearch.class), eq(pageable)))
                 .thenReturn(new PageImpl<>(List.of(course), pageable, 1));
-        when(bakeryImageRepository.findAllByBakeryIdInAndDisplayOrder(anyList(), eq(1)))
-                .thenReturn(List.of());
         when(courseLikeRepository.countByCourseIdIn(List.of(2L))).thenReturn(List.of());
 
         CourseListResponse result =
@@ -199,8 +157,7 @@ class CourseServiceTest {
         when(courseRepository.findByIdAndActiveTrue(7L)).thenReturn(Optional.of(course));
         when(courseBakeryRepository.findAllByCourseIdOrderByVisitOrder(7L))
                 .thenReturn(new ArrayList<>(course.getCourseBakeries()));
-        when(bakeryImageRepository.findAllByBakeryIdInAndDisplayOrder(List.of(10L), 1))
-                .thenReturn(List.of());
+        when(courseSummaryAssembler.buildThumbnailMap(anyList())).thenReturn(Map.of());
         when(courseLikeRepository.countByCourse(course)).thenReturn(0L);
         when(courseLikeRepository.existsByCourseIdAndUserId(7L, 1L)).thenReturn(false);
         when(routeRepository.existsByCourseIdAndUserId(7L, 1L)).thenReturn(true);
@@ -220,8 +177,7 @@ class CourseServiceTest {
         when(courseRepository.findByIdAndActiveTrue(8L)).thenReturn(Optional.of(course));
         when(courseBakeryRepository.findAllByCourseIdOrderByVisitOrder(8L))
                 .thenReturn(new ArrayList<>(course.getCourseBakeries()));
-        when(bakeryImageRepository.findAllByBakeryIdInAndDisplayOrder(List.of(10L), 1))
-                .thenReturn(List.of());
+        when(courseSummaryAssembler.buildThumbnailMap(anyList())).thenReturn(Map.of());
         when(courseLikeRepository.countByCourse(course)).thenReturn(1L);
         when(courseLikeRepository.existsByCourseIdAndUserId(8L, 999L)).thenReturn(false);
         when(routeRepository.existsByCourseIdAndUserId(8L, 999L)).thenReturn(false);
@@ -379,83 +335,6 @@ class CourseServiceTest {
     }
 
     @Test
-    void createAi_returns_job_id_when_async_submit_ok() {
-        AiCourseRequest request = new AiCourseRequest();
-        when(aiCourseAsyncService.processAiCourse(anyString(), eq(2L), eq(request)))
-                .thenReturn(CompletableFuture.completedFuture(null));
-
-        String jobId = courseService.createAi(2L, request);
-
-        assertThat(jobId).isNotBlank();
-        verify(aiCourseRedisService).savePending(eq(jobId), eq(2L));
-        verify(aiCourseAsyncService).processAiCourse(eq(jobId), eq(2L), eq(request));
-    }
-
-    @Test
-    void createAi_throws_whenAsyncSubmitFails() {
-        AiCourseRequest request = new AiCourseRequest();
-        when(aiCourseAsyncService.processAiCourse(anyString(), eq(2L), eq(request)))
-                .thenThrow(new IllegalStateException("submit"));
-
-        assertThatThrownBy(() -> courseService.createAi(2L, request))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.AI_SERVER_ERROR);
-
-        verify(aiCourseRedisService).saveFailed(anyString(), eq("작업 제출에 실패했습니다."));
-    }
-
-    @Test
-    void getAiJobStatus_throws_whenUnknownJob() {
-        when(aiCourseRedisService.findByJobId("job-1", 1L)).thenReturn(Optional.empty());
-
-        assertThatThrownBy(() -> courseService.getAiJobStatus("job-1", 1L))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.AI_JOB_NOT_FOUND);
-    }
-
-    @Test
-    void getAiJobStatus_returns_cached_when_job_exists() {
-        AiJobStatusResponse expected = new AiJobStatusResponse(AiJobStatus.COMPLETED, null);
-        when(aiCourseRedisService.findByJobId("job-2", 1L)).thenReturn(Optional.of(expected));
-
-        assertThat(courseService.getAiJobStatus("job-2", 1L)).isSameAs(expected);
-    }
-
-    @Test
-    void deleteAi_throws_whenManualCourse() {
-        Course course = manualCourse(1L, "수동");
-        when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-
-        assertThatThrownBy(() -> courseService.deleteAi(1L, 1L))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.NOT_AI_COURSE);
-    }
-
-    @Test
-    void deleteAi_throws_whenNotOwner() {
-        Course course = aiPrivateCourse(2L, owner(1L));
-        when(courseRepository.findByIdAndActiveTrue(2L)).thenReturn(Optional.of(course));
-
-        assertThatThrownBy(() -> courseService.deleteAi(2L, 99L))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.FORBIDDEN);
-    }
-
-    @Test
-    void deleteAi_deletes_whenOwner() {
-        Course course = aiPrivateCourse(2L, owner(5L));
-        when(courseRepository.findByIdAndActiveTrue(2L)).thenReturn(Optional.of(course));
-
-        courseService.deleteAi(2L, 5L);
-
-        assertThat(course.isActive()).isFalse();
-    }
-
-    @Test
     void like_throws_whenPrivateAndNotOwner() {
         Course course = aiPrivateCourse(1L, owner(1L));
         when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
@@ -469,8 +348,12 @@ class CourseServiceTest {
     @Test
     void like_throws_whenAlreadyLiked() {
         Course course = manualCourse(1L, "공유");
+        User user = user(3L);
         when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-        when(courseLikeRepository.existsByCourseIdAndUserId(1L, 3L)).thenReturn(true);
+        when(userRepository.findById(3L)).thenReturn(Optional.of(user));
+        doThrow(new DataIntegrityViolationException("dup"))
+                .when(courseLikeRepository)
+                .saveAndFlush(any(CourseLike.class));
 
         assertThatThrownBy(() -> courseService.like(1L, 3L))
                 .isInstanceOf(CustomException.class)
@@ -483,13 +366,12 @@ class CourseServiceTest {
         Course course = manualCourse(1L, "공유");
         User user = user(3L);
         when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-        when(courseLikeRepository.existsByCourseIdAndUserId(1L, 3L)).thenReturn(false);
         when(userRepository.findById(3L)).thenReturn(Optional.of(user));
 
         courseService.like(1L, 3L);
 
         ArgumentCaptor<CourseLike> captor = ArgumentCaptor.forClass(CourseLike.class);
-        verify(courseLikeRepository).save(captor.capture());
+        verify(courseLikeRepository).saveAndFlush(captor.capture());
         assertThat(captor.getValue().getCourse()).isSameAs(course);
         assertThat(captor.getValue().getUser()).isSameAs(user);
     }
@@ -499,11 +381,10 @@ class CourseServiceTest {
         Course course = manualCourse(1L, "공유");
         User user = user(3L);
         when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-        when(courseLikeRepository.existsByCourseIdAndUserId(1L, 3L)).thenReturn(false);
         when(userRepository.findById(3L)).thenReturn(Optional.of(user));
         doThrow(new DataIntegrityViolationException("dup"))
                 .when(courseLikeRepository)
-                .save(any(CourseLike.class));
+                .saveAndFlush(any(CourseLike.class));
 
         assertThatThrownBy(() -> courseService.like(1L, 3L))
                 .isInstanceOf(CustomException.class)
@@ -515,7 +396,6 @@ class CourseServiceTest {
     void like_throws_when_user_missing() {
         Course course = manualCourse(1L, "shared-course");
         when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-        when(courseLikeRepository.existsByCourseIdAndUserId(1L, 3L)).thenReturn(false);
         when(userRepository.findById(3L)).thenReturn(Optional.empty());
 
         assertThatThrownBy(() -> courseService.like(1L, 3L))
@@ -564,8 +444,12 @@ class CourseServiceTest {
     @Test
     void saveRoute_throws_whenAlreadySaved() {
         Course course = manualCourse(1L, "공유");
+        User user = user(4L);
         when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-        when(routeRepository.existsByCourseIdAndUserId(1L, 4L)).thenReturn(true);
+        when(userRepository.findById(4L)).thenReturn(Optional.of(user));
+        doThrow(new DataIntegrityViolationException("dup"))
+                .when(routeRepository)
+                .saveAndFlush(any(Route.class));
 
         assertThatThrownBy(() -> courseService.saveRoute(1L, 4L))
                 .isInstanceOf(CustomException.class)
@@ -578,13 +462,12 @@ class CourseServiceTest {
         Course course = manualCourse(1L, "공유");
         User user = user(4L);
         when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-        when(routeRepository.existsByCourseIdAndUserId(1L, 4L)).thenReturn(false);
         when(userRepository.findById(4L)).thenReturn(Optional.of(user));
 
         courseService.saveRoute(1L, 4L);
 
         ArgumentCaptor<Route> captor = ArgumentCaptor.forClass(Route.class);
-        verify(routeRepository).save(captor.capture());
+        verify(routeRepository).saveAndFlush(captor.capture());
         assertThat(captor.getValue().getCourse()).isSameAs(course);
         assertThat(captor.getValue().getUser()).isSameAs(user);
     }
@@ -594,11 +477,10 @@ class CourseServiceTest {
         Course course = manualCourse(1L, "공유");
         User user = user(4L);
         when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-        when(routeRepository.existsByCourseIdAndUserId(1L, 4L)).thenReturn(false);
         when(userRepository.findById(4L)).thenReturn(Optional.of(user));
         doThrow(new DataIntegrityViolationException("dup"))
                 .when(routeRepository)
-                .save(any(Route.class));
+                .saveAndFlush(any(Route.class));
 
         assertThatThrownBy(() -> courseService.saveRoute(1L, 4L))
                 .isInstanceOf(CustomException.class)
@@ -610,7 +492,6 @@ class CourseServiceTest {
     void saveRoute_throws_when_user_missing() {
         Course course = manualCourse(1L, "shared-course");
         when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-        when(routeRepository.existsByCourseIdAndUserId(1L, 4L)).thenReturn(false);
         when(userRepository.findById(4L)).thenReturn(Optional.empty());
 
         assertThatThrownBy(() -> courseService.saveRoute(1L, 4L))
@@ -650,582 +531,6 @@ class CourseServiceTest {
         courseService.removeRoute(1L, 2L);
 
         verify(routeRepository).delete(route);
-    }
-
-    // ── getDrivingRoute ──────────────────────────────────────────────────
-
-    @Test
-    void getDrivingRoute_throws_whenCourseNotFound() {
-        when(courseRepository.findActiveWithBakeriesById(1L)).thenReturn(Optional.empty());
-
-        assertThatThrownBy(() -> courseService.getDrivingRoute(1L, 1L, UserRole.ROLE_USER))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.COURSE_NOT_FOUND);
-    }
-
-    @Test
-    void getDrivingRoute_throws_unauthorized_whenPrivateCourseAndAnonymous() {
-        Course course = aiPrivateCourse(1L, owner(1L));
-        when(courseRepository.findActiveWithBakeriesById(1L)).thenReturn(Optional.of(course));
-
-        assertThatThrownBy(() -> courseService.getDrivingRoute(1L, null, null))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.UNAUTHORIZED);
-    }
-
-    @Test
-    void getDrivingRoute_throws_forbidden_whenPrivateCourseAndNotOwner() {
-        Course course = aiPrivateCourse(1L, owner(1L));
-        when(courseRepository.findActiveWithBakeriesById(1L)).thenReturn(Optional.of(course));
-
-        assertThatThrownBy(() -> courseService.getDrivingRoute(1L, 2L, UserRole.ROLE_USER))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.FORBIDDEN);
-    }
-
-    @Test
-    void getDrivingRoute_allows_admin_to_access_private_course() {
-        Course course = aiPrivateCourse(1L, owner(1L));
-        List<Coordinate> cachedPath =
-                List.of(new Coordinate(36.0, 127.0), new Coordinate(36.1, 127.1));
-        CourseDrivingRoute cached =
-                CourseDrivingRoute.builder().courseId(1L).path(cachedPath).build();
-
-        when(courseRepository.findActiveWithBakeriesById(1L)).thenReturn(Optional.of(course));
-        when(courseDrivingRouteRepository.findById(1L)).thenReturn(Optional.of(cached));
-
-        DrivingRouteResponse response =
-                courseService.getDrivingRoute(1L, 999L, UserRole.ROLE_ADMIN);
-
-        assertThat(response.getPath()).isEqualTo(cachedPath);
-        verify(drivingRouteClient, never()).getPath(any());
-    }
-
-    @Test
-    void getDrivingRoute_returns_cached_path_whenCacheHit() {
-        Course course = manualCourse(1L, "공유코스");
-        List<Coordinate> cachedPath =
-                List.of(new Coordinate(36.0, 127.0), new Coordinate(36.1, 127.1));
-        CourseDrivingRoute cached =
-                CourseDrivingRoute.builder().courseId(1L).path(cachedPath).build();
-
-        when(courseRepository.findActiveWithBakeriesById(1L)).thenReturn(Optional.of(course));
-        when(courseDrivingRouteRepository.findById(1L)).thenReturn(Optional.of(cached));
-
-        DrivingRouteResponse response = courseService.getDrivingRoute(1L, null, null);
-
-        assertThat(response.getPath()).isEqualTo(cachedPath);
-        verify(drivingRouteClient, never()).getPath(any());
-    }
-
-    @Test
-    @SuppressWarnings("unchecked")
-    void getDrivingRoute_manual_fetchesAndSaves_whenCacheMiss() {
-        Course course = manualCourse(1L, "공유코스");
-        Bakery b1 = bakeryAt(10L, "A빵집", 36.0, 127.0);
-        Bakery b2 = bakeryAt(20L, "B빵집", 36.1, 127.1);
-        course.addCourseBakery(CourseBakery.builder().visitOrder(1).bakery(b1).build());
-        course.addCourseBakery(CourseBakery.builder().visitOrder(2).bakery(b2).build());
-
-        List<Coordinate> path = List.of(new Coordinate(36.0, 127.0), new Coordinate(36.1, 127.1));
-        RouteResult routeResult = new RouteResult(path, List.of(), 600);
-
-        when(courseRepository.findActiveWithBakeriesById(1L)).thenReturn(Optional.of(course));
-        when(courseDrivingRouteRepository.findById(1L)).thenReturn(Optional.empty());
-        when(drivingRouteClient.getPath(any())).thenReturn(routeResult);
-
-        DrivingRouteResponse response = courseService.getDrivingRoute(1L, null, null);
-
-        assertThat(response.getPath()).isEqualTo(path);
-
-        // MANUAL 코스는 사용자 위치 미포함 — 빵집만 좌표로 전달
-        ArgumentCaptor<List<Coordinate>> coordCaptor = ArgumentCaptor.forClass(List.class);
-        verify(drivingRouteClient).getPath(coordCaptor.capture());
-        List<Coordinate> coords = coordCaptor.getValue();
-        assertThat(coords).hasSize(2);
-        assertThat(coords.get(0).getLat()).isEqualTo(36.0);
-        assertThat(coords.get(0).getLng()).isEqualTo(127.0);
-
-        verify(courseDrivingRouteSaver).save(eq(1L), any(RouteResult.class));
-    }
-
-    @Test
-    @SuppressWarnings("unchecked")
-    void getDrivingRoute_ai_prependsUserLocation_whenCacheMiss() {
-        User owner = owner(1L);
-        Course course = aiCourseWithLocation(2L, owner, 35.5, 126.5);
-        Bakery bakery = bakeryAt(10L, "빵집", 36.0, 127.0);
-        course.addCourseBakery(CourseBakery.builder().visitOrder(1).bakery(bakery).build());
-
-        List<Coordinate> path = List.of(new Coordinate(35.5, 126.5), new Coordinate(36.0, 127.0));
-        RouteResult routeResult = new RouteResult(path, List.of(), 600);
-
-        when(courseRepository.findActiveWithBakeriesById(2L)).thenReturn(Optional.of(course));
-        when(courseDrivingRouteRepository.findById(2L)).thenReturn(Optional.empty());
-        when(drivingRouteClient.getPath(any())).thenReturn(routeResult);
-
-        courseService.getDrivingRoute(2L, 1L, UserRole.ROLE_USER);
-
-        // AI 코스는 AiCourseInfo의 유저 위치가 첫 번째 좌표로 추가되어야 함
-        ArgumentCaptor<List<Coordinate>> coordCaptor = ArgumentCaptor.forClass(List.class);
-        verify(drivingRouteClient).getPath(coordCaptor.capture());
-        List<Coordinate> coords = coordCaptor.getValue();
-        assertThat(coords).hasSize(2); // 유저위치 + 빵집 1개
-        assertThat(coords.get(0).getLat()).isEqualTo(35.5);
-        assertThat(coords.get(0).getLng()).isEqualTo(126.5);
-        assertThat(coords.get(1).getLat()).isEqualTo(36.0);
-    }
-
-    @Test
-    void getDrivingRoute_ai_succeeds_withOneBakery() {
-        User owner = owner(1L);
-        Course course = aiCourseWithLocation(2L, owner, 35.5, 126.5);
-        Bakery bakery = bakeryAt(10L, "빵집", 36.0, 127.0);
-        course.addCourseBakery(CourseBakery.builder().visitOrder(1).bakery(bakery).build());
-
-        List<Coordinate> path = List.of(new Coordinate(35.5, 126.5), new Coordinate(36.0, 127.0));
-        RouteResult routeResult = new RouteResult(path, List.of(), 600);
-
-        when(courseRepository.findActiveWithBakeriesById(2L)).thenReturn(Optional.of(course));
-        when(courseDrivingRouteRepository.findById(2L)).thenReturn(Optional.empty());
-        when(drivingRouteClient.getPath(any())).thenReturn(routeResult);
-
-        DrivingRouteResponse response = courseService.getDrivingRoute(2L, 1L, UserRole.ROLE_USER);
-
-        assertThat(response.getPath()).isEqualTo(path);
-        verify(drivingRouteClient).getPath(any());
-    }
-
-    @Test
-    void getDrivingRoute_throws_whenManualCourseHasOnlyOneBakery() {
-        Course course = manualCourse(1L, "공유코스");
-        Bakery bakery = bakeryAt(10L, "빵집", 36.0, 127.0);
-        course.addCourseBakery(CourseBakery.builder().visitOrder(1).bakery(bakery).build());
-
-        when(courseRepository.findActiveWithBakeriesById(1L)).thenReturn(Optional.of(course));
-        when(courseDrivingRouteRepository.findById(1L)).thenReturn(Optional.empty());
-
-        assertThatThrownBy(() -> courseService.getDrivingRoute(1L, null, null))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.ROUTE_INSUFFICIENT_WAYPOINTS);
-
-        verify(drivingRouteClient, never()).getPath(any());
-    }
-
-    @Test
-    void getDrivingRoute_throws_whenManualCourseExceedsWaypointLimit() {
-        // 수동 코스: 빵집 8개 → coordinates 8개 > 7 (Kakao 최대: origin + 5 waypoints + destination)
-        Course course = manualCourse(1L, "공유코스");
-        for (int i = 0; i < 8; i++) {
-            course.addCourseBakery(
-                    CourseBakery.builder()
-                            .visitOrder(i + 1)
-                            .bakery(bakeryAt(10L + i, "빵집" + i, 36.0 + i * 0.01, 127.0 + i * 0.01))
-                            .build());
-        }
-
-        when(courseRepository.findActiveWithBakeriesById(1L)).thenReturn(Optional.of(course));
-        when(courseDrivingRouteRepository.findById(1L)).thenReturn(Optional.empty());
-
-        assertThatThrownBy(() -> courseService.getDrivingRoute(1L, null, null))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.ROUTE_TOO_MANY_WAYPOINTS);
-
-        verify(drivingRouteClient, never()).getPath(any());
-    }
-
-    @Test
-    void getDrivingRoute_throws_whenAiCourseExceedsWaypointLimit() {
-        // AI 코스: 빵집 7개 + 출발 위치 1개 → coordinates 8개 > 7
-        User owner = owner(1L);
-        Course course = aiCourseWithLocation(2L, owner, 35.5, 126.5);
-        for (int i = 0; i < 7; i++) {
-            course.addCourseBakery(
-                    CourseBakery.builder()
-                            .visitOrder(i + 1)
-                            .bakery(bakeryAt(10L + i, "빵집" + i, 36.0 + i * 0.01, 127.0 + i * 0.01))
-                            .build());
-        }
-
-        when(courseRepository.findActiveWithBakeriesById(2L)).thenReturn(Optional.of(course));
-        when(courseDrivingRouteRepository.findById(2L)).thenReturn(Optional.empty());
-
-        assertThatThrownBy(() -> courseService.getDrivingRoute(2L, 1L, UserRole.ROLE_USER))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.ROUTE_TOO_MANY_WAYPOINTS);
-
-        verify(drivingRouteClient, never()).getPath(any());
-    }
-
-    @Test
-    void getDrivingRoute_throws_whenAiCourseHasNoAiInfo() {
-        User owner = owner(1L);
-        Course course = aiPrivateCourse(1L, owner);
-        ReflectionTestUtils.setField(course, "aiCourseInfo", null); // aiInfo null 강제
-        Bakery bakery = bakeryAt(10L, "빵집", 36.0, 127.0);
-        course.addCourseBakery(CourseBakery.builder().visitOrder(1).bakery(bakery).build());
-
-        when(courseRepository.findActiveWithBakeriesById(1L)).thenReturn(Optional.of(course));
-        when(courseDrivingRouteRepository.findById(1L)).thenReturn(Optional.empty());
-
-        assertThatThrownBy(() -> courseService.getDrivingRoute(1L, 1L, UserRole.ROLE_USER))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.ROUTE_NOT_FOUND);
-
-        verify(drivingRouteClient, never()).getPath(any());
-    }
-
-    @Test
-    void getDrivingRoute_ignores_concurrent_save_conflict() {
-        Course course = manualCourse(1L, "공유코스");
-        Bakery b1 = bakeryAt(10L, "A빵집", 36.0, 127.0);
-        Bakery b2 = bakeryAt(20L, "B빵집", 36.1, 127.1);
-        course.addCourseBakery(CourseBakery.builder().visitOrder(1).bakery(b1).build());
-        course.addCourseBakery(CourseBakery.builder().visitOrder(2).bakery(b2).build());
-
-        List<Coordinate> path = List.of(new Coordinate(36.0, 127.0), new Coordinate(36.1, 127.1));
-        RouteResult routeResult = new RouteResult(path, List.of(), 600);
-
-        when(courseRepository.findActiveWithBakeriesById(1L)).thenReturn(Optional.of(course));
-        when(courseDrivingRouteRepository.findById(1L)).thenReturn(Optional.empty());
-        when(drivingRouteClient.getPath(any())).thenReturn(routeResult);
-        doThrow(new DataIntegrityViolationException("dup"))
-                .when(courseDrivingRouteSaver)
-                .save(eq(1L), any(RouteResult.class));
-
-        // PK 충돌이 나도 경로는 정상 반환
-        DrivingRouteResponse response = courseService.getDrivingRoute(1L, null, null);
-
-        assertThat(response.getPath()).isEqualTo(path);
-    }
-
-    // ── reorderBakeries ──────────────────────────────────────────────────
-
-    @Test
-    void reorderBakeries_throws_whenCourseNotFound() {
-        when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.empty());
-
-        assertThatThrownBy(
-                        () ->
-                                courseService.reorderBakeries(
-                                        1L, 1L, UserRole.ROLE_ADMIN, reorderRequest(List.of(10L))))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.COURSE_NOT_FOUND);
-    }
-
-    @Test
-    void reorderBakeries_throws_whenForbidden() {
-        Course course = manualCourse(1L, "코스"); // user == null
-        when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-
-        assertThatThrownBy(
-                        () ->
-                                courseService.reorderBakeries(
-                                        1L, 99L, UserRole.ROLE_USER, reorderRequest(List.of(10L))))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.FORBIDDEN);
-    }
-
-    @Test
-    void reorderBakeries_throws_whenBakeryOrderEmpty() {
-        Course course = manualCourse(1L, "코스");
-        when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-
-        assertThatThrownBy(
-                        () ->
-                                courseService.reorderBakeries(
-                                        1L, 99L, UserRole.ROLE_ADMIN, reorderRequest(List.of())))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.INVALID_INPUT_VALUE);
-    }
-
-    @Test
-    void reorderBakeries_throws_whenDuplicateIds() {
-        Course course = manualCourse(1L, "코스");
-        Bakery b1 = bakeryAt(10L, "A", 36.0, 127.0);
-        CourseBakery cb1 = CourseBakery.builder().visitOrder(1).bakery(b1).build();
-        when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-        when(courseBakeryRepository.findAllByCourseId(1L)).thenReturn(List.of(cb1));
-
-        assertThatThrownBy(
-                        () ->
-                                courseService.reorderBakeries(
-                                        1L,
-                                        99L,
-                                        UserRole.ROLE_ADMIN,
-                                        reorderRequest(List.of(10L, 10L))))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.INVALID_INPUT_VALUE);
-    }
-
-    @Test
-    void reorderBakeries_throws_whenCountMismatch() {
-        Course course = manualCourse(1L, "코스");
-        Bakery b1 = bakeryAt(10L, "A", 36.0, 127.0);
-        Bakery b2 = bakeryAt(20L, "B", 36.1, 127.1);
-        when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-        when(courseBakeryRepository.findAllByCourseId(1L))
-                .thenReturn(
-                        List.of(
-                                CourseBakery.builder().visitOrder(1).bakery(b1).build(),
-                                CourseBakery.builder().visitOrder(2).bakery(b2).build()));
-
-        // 활성 빵집은 2개인데 1개만 보냄
-        assertThatThrownBy(
-                        () ->
-                                courseService.reorderBakeries(
-                                        1L, 99L, UserRole.ROLE_ADMIN, reorderRequest(List.of(10L))))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.BAKERY_ORDER_COUNT_MISMATCH);
-    }
-
-    @Test
-    void reorderBakeries_updatesVisitOrder_andReturnsResponse() {
-        Course course = manualCourse(1L, "코스");
-        Bakery b1 = bakeryAt(10L, "A", 36.0, 127.0);
-        Bakery b2 = bakeryAt(20L, "B", 36.1, 127.1);
-        CourseBakery cb1 = CourseBakery.builder().visitOrder(1).bakery(b1).build();
-        CourseBakery cb2 = CourseBakery.builder().visitOrder(2).bakery(b2).build();
-
-        when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-        when(courseBakeryRepository.findAllByCourseId(1L)).thenReturn(List.of(cb1, cb2));
-
-        List<Coordinate> path = List.of(new Coordinate(36.1, 127.1), new Coordinate(36.0, 127.0));
-        when(drivingRouteClient.getPath(any())).thenReturn(new RouteResult(path, List.of(), 600));
-
-        // 순서 뒤집기: [20, 10]
-        ReorderBakeriesResponse response =
-                courseService.reorderBakeries(
-                        1L, 99L, UserRole.ROLE_ADMIN, reorderRequest(List.of(20L, 10L)));
-
-        assertThat(response.getCourseId()).isEqualTo(1L);
-        assertThat(response.getBakeryOrder()).containsExactly(20L, 10L);
-        assertThat(cb2.getVisitOrder()).isEqualTo(1); // 20L → 1순위
-        assertThat(cb1.getVisitOrder()).isEqualTo(2); // 10L → 2순위
-        verify(courseDrivingRouteRepository).deleteAllByCourseIdIn(List.of(1L));
-    }
-
-    @Test
-    void reorderBakeries_filtersInactiveIds_andProceedsWhenActiveSetMatches() {
-        Course course = manualCourse(1L, "코스");
-        Bakery b1 = bakeryAt(10L, "A", 36.0, 127.0);
-        Bakery b2 = bakeryAt(20L, "B", 36.1, 127.1);
-        CourseBakery cb1 = CourseBakery.builder().visitOrder(1).bakery(b1).build();
-        CourseBakery cb2 = CourseBakery.builder().visitOrder(2).bakery(b2).build();
-
-        when(courseRepository.findByIdAndActiveTrue(1L)).thenReturn(Optional.of(course));
-        when(courseBakeryRepository.findAllByCourseId(1L)).thenReturn(List.of(cb1, cb2));
-        when(drivingRouteClient.getPath(any()))
-                .thenReturn(
-                        new RouteResult(
-                                List.of(new Coordinate(36.0, 127.0), new Coordinate(36.1, 127.1)),
-                                List.of(),
-                                600));
-
-        // 비활성 ID 999는 필터링 후 활성 [10, 20]과 일치 → 성공
-        ReorderBakeriesResponse response =
-                courseService.reorderBakeries(
-                        1L, 99L, UserRole.ROLE_ADMIN, reorderRequest(List.of(999L, 10L, 20L)));
-
-        assertThat(response.getBakeryOrder()).containsExactly(10L, 20L);
-    }
-
-    // ── updateManual 추가 케이스 ─────────────────────────────────────────
-
-    // ── getAiPreview ─────────────────────────────────────────────────────
-
-    // ── getAiPreview ──────────────────────────────────────────────────────
-
-    @Test
-    void getAiPreview_throws_whenResultNotFound() {
-        when(aiCourseResultRedisService.getResult("job-p1", 1L)).thenReturn(Optional.empty());
-
-        assertThatThrownBy(() -> courseService.getAiPreview("job-p1", 1L))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.AI_RESULT_NOT_FOUND);
-    }
-
-    @Test
-    void getAiPreview_throws_whenRecommendedBakeryMissing() {
-        // 프리뷰와 저장의 불일치 방지 — 추천 빵집이 DB에 없으면 프리뷰도 예외
-        AiCourseResultCache cache = resultCache(1L, 10L);
-        when(aiCourseResultRedisService.getResult("job-p2", 1L)).thenReturn(Optional.of(cache));
-        when(bakeryRepository.findAllByIdInAndActiveTrue(anyList())).thenReturn(List.of());
-
-        assertThatThrownBy(() -> courseService.getAiPreview("job-p2", 1L))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.AI_RECOMMENDED_BAKERY_NOT_FOUND);
-    }
-
-    @Test
-    void getAiPreview_returns_enriched_preview() {
-        Bakery bakery = bakery(10L, "맛집");
-        AiCourseResultCache cache = resultCache(1L, 10L);
-        when(aiCourseResultRedisService.getResult("job-p3", 1L)).thenReturn(Optional.of(cache));
-        when(bakeryRepository.findAllByIdInAndActiveTrue(anyList())).thenReturn(List.of(bakery));
-
-        AiCoursePreviewResponse preview = courseService.getAiPreview("job-p3", 1L);
-
-        assertThat(preview.getName()).isEqualTo("ai-course-name");
-        assertThat(preview.getBakeryCount()).isEqualTo(1);
-        assertThat(preview.getBakeries()).hasSize(1);
-
-        var b = preview.getBakeries().get(0);
-        assertThat(b.getId()).isEqualTo(10L);
-        assertThat(b.getName()).isEqualTo("맛집"); // DB 이름 사용
-        assertThat(b.getAddress()).isEqualTo("addr");
-        assertThat(b.getRating()).isEqualTo(4.0);
-    }
-
-    // ── saveAiCourse ──────────────────────────────────────────────────────
-
-    @Test
-    void saveAiCourse_throws_whenLockAlreadyHeld() {
-        when(aiCourseResultRedisService.tryAcquireSaveLock("job-s1")).thenReturn(false);
-
-        assertThatThrownBy(() -> courseService.saveAiCourse("job-s1", 1L, null))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.AI_COURSE_SAVE_IN_PROGRESS);
-    }
-
-    @Test
-    void saveAiCourse_throws_whenResultNotFound() {
-        when(aiCourseResultRedisService.tryAcquireSaveLock("job-s2")).thenReturn(true);
-        when(aiCourseResultRedisService.getResult("job-s2", 1L)).thenReturn(Optional.empty());
-
-        assertThatThrownBy(() -> courseService.saveAiCourse("job-s2", 1L, null))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.AI_RESULT_NOT_FOUND);
-    }
-
-    @Test
-    void saveAiCourse_throws_whenRequesterIsNotOwner() {
-        // 비소유자 호출 시 FORBIDDEN — Redis 키 보존
-        when(aiCourseResultRedisService.tryAcquireSaveLock("job-s3")).thenReturn(true);
-        when(aiCourseResultRedisService.getResult("job-s3", 99L))
-                .thenThrow(new CustomException(ErrorCode.FORBIDDEN));
-
-        assertThatThrownBy(() -> courseService.saveAiCourse("job-s3", 99L, null))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.FORBIDDEN);
-    }
-
-    @Test
-    void saveAiCourse_throws_whenRecommendedBakeryMissing() {
-        // 추천 빵집이 DB에 없으면 예외 — Redis 결과는 삭제되지 않음
-        AiCourseResultCache cache = resultCache(1L, 10L);
-        when(aiCourseResultRedisService.tryAcquireSaveLock("job-s4")).thenReturn(true);
-        when(aiCourseResultRedisService.getResult("job-s4", 1L)).thenReturn(Optional.of(cache));
-        when(userRepository.findById(1L)).thenReturn(Optional.of(user(1L)));
-        when(userPreferenceRepository.findByUserId(1L))
-                .thenReturn(Optional.of(preference(user(1L))));
-        when(bakeryRepository.findAllByIdInAndActiveTrue(anyList())).thenReturn(List.of());
-
-        assertThatThrownBy(() -> courseService.saveAiCourse("job-s4", 1L, null))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.AI_RECOMMENDED_BAKERY_NOT_FOUND);
-
-        verify(aiCourseResultRedisService, never()).deleteResult(any());
-    }
-
-    @Test
-    void saveAiCourse_saves_and_returns_courseId() {
-        Bakery bakery = bakery(10L, "맛집");
-        AiCourseResultCache cache = resultCache(1L, 10L);
-
-        when(aiCourseResultRedisService.tryAcquireSaveLock("job-s5")).thenReturn(true);
-        when(aiCourseResultRedisService.getResult("job-s5", 1L)).thenReturn(Optional.of(cache));
-        when(userRepository.findById(1L)).thenReturn(Optional.of(user(1L)));
-        when(userPreferenceRepository.findByUserId(1L))
-                .thenReturn(Optional.of(preference(user(1L))));
-        when(bakeryRepository.findAllByIdInAndActiveTrue(anyList())).thenReturn(List.of(bakery));
-        when(courseRepository.save(any(Course.class)))
-                .thenAnswer(
-                        inv -> {
-                            Course c = inv.getArgument(0);
-                            ReflectionTestUtils.setField(c, "id", 99L);
-                            return c;
-                        });
-
-        Long courseId = courseService.saveAiCourse("job-s5", 1L, null);
-
-        assertThat(courseId).isEqualTo(99L);
-        verify(courseRepository).save(any(Course.class));
-        // 단위 테스트는 트랜잭션 동기화 비활성 → afterCommit 없이 즉시 정리됨
-        verify(aiCourseResultRedisService).deleteResult("job-s5");
-        verify(aiCourseResultRedisService).releaseSaveLock("job-s5");
-        verify(aiCourseRedisService).deleteJob("job-s5");
-    }
-
-    @Test
-    void saveAiCourse_applies_custom_bakeryOrder_whenProvided() {
-        // bakery 10→1번, 20→2번 추천 — 사용자가 [20, 10]으로 역순 지정
-        Bakery b10 = bakery(10L, "빵집10");
-        Bakery b20 = bakery(20L, "빵집20");
-        AiCourseResultCache cache = resultCache2(1L, 10L, 20L);
-
-        when(aiCourseResultRedisService.tryAcquireSaveLock("job-s6")).thenReturn(true);
-        when(aiCourseResultRedisService.getResult("job-s6", 1L)).thenReturn(Optional.of(cache));
-        when(userRepository.findById(1L)).thenReturn(Optional.of(user(1L)));
-        when(userPreferenceRepository.findByUserId(1L))
-                .thenReturn(Optional.of(preference(user(1L))));
-        when(bakeryRepository.findAllByIdInAndActiveTrue(anyList())).thenReturn(List.of(b10, b20));
-
-        ArgumentCaptor<Course> captor = ArgumentCaptor.forClass(Course.class);
-        when(courseRepository.save(captor.capture()))
-                .thenAnswer(
-                        inv -> {
-                            Course c = inv.getArgument(0);
-                            ReflectionTestUtils.setField(c, "id", 100L);
-                            return c;
-                        });
-
-        Long courseId = courseService.saveAiCourse("job-s6", 1L, List.of(20L, 10L));
-
-        assertThat(courseId).isEqualTo(100L);
-        List<CourseBakery> cbs =
-                captor.getValue().getCourseBakeries().stream()
-                        .sorted(java.util.Comparator.comparingInt(CourseBakery::getVisitOrder))
-                        .toList();
-        assertThat(cbs.get(0).getBakery().getId()).isEqualTo(20L); // visitOrder=1
-        assertThat(cbs.get(1).getBakery().getId()).isEqualTo(10L); // visitOrder=2
-    }
-
-    @Test
-    void saveAiCourse_throws_whenBakeryOrderMismatch() {
-        // 추천 빵집은 10, 20인데 잘못된 ID(99)가 포함 → 400
-        AiCourseResultCache cache = resultCache2(1L, 10L, 20L);
-
-        when(aiCourseResultRedisService.tryAcquireSaveLock("job-s7")).thenReturn(true);
-        when(aiCourseResultRedisService.getResult("job-s7", 1L)).thenReturn(Optional.of(cache));
-        when(userRepository.findById(1L)).thenReturn(Optional.of(user(1L)));
-        when(userPreferenceRepository.findByUserId(1L))
-                .thenReturn(Optional.of(preference(user(1L))));
-        when(bakeryRepository.findAllByIdInAndActiveTrue(anyList()))
-                .thenReturn(List.of(bakery(10L, "빵집10"), bakery(20L, "빵집20")));
-
-        assertThatThrownBy(() -> courseService.saveAiCourse("job-s7", 1L, List.of(10L, 99L)))
-                .isInstanceOf(CustomException.class)
-                .extracting("errorCode")
-                .isEqualTo(ErrorCode.BAKERY_ORDER_COUNT_MISMATCH);
     }
 
     @Test
@@ -1322,131 +627,5 @@ class CourseServiceTest {
         ReflectionTestUtils.setField(request, "name", name);
         ReflectionTestUtils.setField(request, "bakeryIds", bakeryIds);
         return request;
-    }
-
-    private static Bakery bakeryAt(long id, String name, double lat, double lng) {
-        Bakery b =
-                Bakery.builder()
-                        .name(name)
-                        .address("addr")
-                        .region("서울")
-                        .latitude(lat)
-                        .longitude(lng)
-                        .phone("010")
-                        .rating(4.0)
-                        .mapLink("m")
-                        .dineInAvailable(true)
-                        .parkingAvailable(false)
-                        .drinkAvailable(true)
-                        .note("")
-                        .build();
-        ReflectionTestUtils.setField(b, "id", id);
-        return b;
-    }
-
-    private static AiCourseRequest aiRequest() {
-        AiCourseRequest req = new AiCourseRequest();
-        ReflectionTestUtils.setField(req, "travelType", TravelType.ALONE);
-        ReflectionTestUtils.setField(req, "budgetRange", BudgetRange.ANY);
-        ReflectionTestUtils.setField(req, "minimizeRoute", false);
-        ReflectionTestUtils.setField(req, "breadTypes", List.of(BreadType.BREAD));
-        ReflectionTestUtils.setField(req, "latitude", 36.0);
-        ReflectionTestUtils.setField(req, "longitude", 127.0);
-        ReflectionTestUtils.setField(req, "waitingPreference", false);
-        ReflectionTestUtils.setField(req, "drinkPreference", false);
-        ReflectionTestUtils.setField(req, "bakeryCount", 2);
-        ReflectionTestUtils.setField(req, "flexibilityLevel", FlexibilityLevel.ACTIVE);
-        return req;
-    }
-
-    private static AiCourseWebhookResponse webhookResponse(long bakeryId) {
-        RecommendedBakeryResponse rb = new RecommendedBakeryResponse();
-        ReflectionTestUtils.setField(rb, "id", bakeryId);
-        ReflectionTestUtils.setField(rb, "order", 1);
-        ReflectionTestUtils.setField(rb, "recommendedBread", "소금빵");
-        ReflectionTestUtils.setField(rb, "reason", "맛있음");
-
-        AiCourseWebhookResponse res = new AiCourseWebhookResponse();
-        ReflectionTestUtils.setField(res, "name", "ai-course-name");
-        ReflectionTestUtils.setField(res, "theme", "테마");
-        ReflectionTestUtils.setField(res, "estimatedCost", 8000L);
-        ReflectionTestUtils.setField(res, "estimatedTime", "2h");
-        ReflectionTestUtils.setField(res, "summary", "요약");
-        ReflectionTestUtils.setField(res, "recommendReason", "추천");
-        ReflectionTestUtils.setField(res, "bakeries", List.of(rb));
-        return res;
-    }
-
-    private static AiCourseResultCache resultCache(long userId, long bakeryId) {
-        return AiCourseResultCache.builder()
-                .userId(userId)
-                .request(aiRequest())
-                .response(webhookResponse(bakeryId))
-                .build();
-    }
-
-    /** 빵집 2개짜리 결과 캐시 (순서 지정 테스트용) */
-    private static AiCourseResultCache resultCache2(long userId, long bakeryId1, long bakeryId2) {
-        RecommendedBakeryResponse rb1 = new RecommendedBakeryResponse();
-        ReflectionTestUtils.setField(rb1, "id", bakeryId1);
-        ReflectionTestUtils.setField(rb1, "order", 1);
-        ReflectionTestUtils.setField(rb1, "recommendedBread", "크루아상");
-        ReflectionTestUtils.setField(rb1, "reason", "이유1");
-
-        RecommendedBakeryResponse rb2 = new RecommendedBakeryResponse();
-        ReflectionTestUtils.setField(rb2, "id", bakeryId2);
-        ReflectionTestUtils.setField(rb2, "order", 2);
-        ReflectionTestUtils.setField(rb2, "recommendedBread", "소금빵");
-        ReflectionTestUtils.setField(rb2, "reason", "이유2");
-
-        AiCourseWebhookResponse res = new AiCourseWebhookResponse();
-        ReflectionTestUtils.setField(res, "name", "ai-course-name");
-        ReflectionTestUtils.setField(res, "theme", "테마");
-        ReflectionTestUtils.setField(res, "estimatedCost", 15000L);
-        ReflectionTestUtils.setField(res, "estimatedTime", "2h");
-        ReflectionTestUtils.setField(res, "summary", "요약");
-        ReflectionTestUtils.setField(res, "recommendReason", "추천");
-        ReflectionTestUtils.setField(res, "bakeries", List.of(rb1, rb2));
-
-        return AiCourseResultCache.builder()
-                .userId(userId)
-                .request(aiRequest())
-                .response(res)
-                .build();
-    }
-
-    private static UserPreference preference(User user) {
-        return UserPreference.builder()
-                .bakeryTypes(List.of(com.breadbread.bakery.entity.BakeryType.CLASSIC))
-                .bakeryPersonalities(List.of())
-                .bakeryUseTypes(List.of())
-                .waitingTolerance(WaitingTolerance.NO_WAIT)
-                .user(user)
-                .build();
-    }
-
-    private static ReorderBakeriesRequest reorderRequest(List<Long> bakeryOrder) {
-        ReorderBakeriesRequest request = new ReorderBakeriesRequest();
-        ReflectionTestUtils.setField(request, "bakeryOrder", bakeryOrder);
-        return request;
-    }
-
-    private static Course aiCourseWithLocation(long id, User owner, double lat, double lng) {
-        AiCourseInfo aiInfo =
-                AiCourseInfo.builder()
-                        .travelType(TravelType.ALONE)
-                        .budgetRange(BudgetRange.ANY)
-                        .waitingPreference(false)
-                        .drinkPreference(false)
-                        .bakeryCount(1)
-                        .flexibilityLevel(FlexibilityLevel.ACTIVE)
-                        .recommendReason("r")
-                        .minimizeRoute(false)
-                        .latitude(lat)
-                        .longitude(lng)
-                        .build();
-        Course course = Course.createAi("AI 코스", owner, null, aiInfo, Set.of());
-        ReflectionTestUtils.setField(course, "id", id);
-        return course;
     }
 }
